[MOV] event(_sms): split models / file
Just finish the split of models into files, with type specific models
being extracted.
