+++ addons/event/models/__init__.py
@@ -1,6 +1,11 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+# some inherit require to define those models first
+from . import event_type
+from . import event_type_mail
+from . import event_type_ticket
+
 from . import event_event
 from . import event_mail
 from . import event_mail_registration
