+++ addons/event_sms/models/__init__.py
@@ -3,4 +3,5 @@
 
 from . import event_mail
 from . import event_mail_registration
+from . import event_type_mail
 from . import sms_template
