+++ addons/event_sms/models/event_mail.py
@@ -1,19 +1,4 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from odoo import fields, models
-
-
-class EventTypeMail(models.Model):
-    _inherit = 'event.type.mail'
-
-    notification_type = fields.Selection(selection_add=[('sms', 'SMS')])
-    template_ref = fields.Reference(ondelete={'sms.template': 'cascade'}, selection_add=[('sms.template', 'SMS')])
-
-    def _compute_notification_type(self):
-        super()._compute_notification_type()
-        sms_schedulers = self.filtered(lambda scheduler: scheduler.template_ref and scheduler.template_ref._name == 'sms.template')
-        sms_schedulers.notification_type = 'sms'
 
 
 class EventMail(models.Model):
